+++ source/aws-connect-vm-serverless/src/main/java/com/amazonaws/helper/MetricsUtil.java
@@ -4,9 +4,12 @@ import org.json.JSONObject;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import java.io.BufferedReader;
+import java.io.BufferedReader; 
+import java.io.IOException; 
+import java.io.InputStream; 
 import java.io.DataOutputStream;
 import java.io.InputStreamReader;
+import java.io.OutputStream; 
 import java.io.Reader;
 import java.net.HttpURLConnection;
 import java.net.URL;
@@ -38,8 +41,12 @@ public final class MetricsUtil {
     private static final Logger logger = LoggerFactory.getLogger(MetricsUtil.class);
 
     public static void sendMetrics(String process, String status, String details, boolean isTranscribed, boolean isEncrypted, String languageCode) {
+        if (!System.getenv("SEND_ANON_DATA").equals("true")) {
+            return;
+        }
+        
         try {
-            Map<String, String> parameters = createParameters(process, status, details, isTranscribed, isEncrypted, languageCode);
+            JSONObject parameters = createParameters(process, status, details, isTranscribed, isEncrypted, languageCode);
             HttpURLConnection connection = createConnection(parameters);
             
             connection.connect();
@@ -53,38 +60,39 @@ public final class MetricsUtil {
         }
     }
 
-    private static Map<String, String> createParameters(String process, String status, String details,
+    private static JSONObject createParameters(String process, String status, String details,
                                                         boolean isTranscribed, boolean isEncrypted, String languageCode) {
-        JSONObject json = new JSONObject();
-        json.put("process", process);
-        json.put("status", status);
-        json.put("details", details);
-        json.put("isTranscribed", String.valueOf(isTranscribed));
-        json.put("isEncrypted", String.valueOf(isEncrypted));
-        json.put("languageCode", languageCode);
+        JSONObject dataJson = new JSONObject();
+        dataJson.put("process", process);
+        dataJson.put("status", status);
+        dataJson.put("details", details);
+        dataJson.put("isTranscribed", String.valueOf(isTranscribed));
+        dataJson.put("isEncrypted", String.valueOf(isEncrypted));
+        dataJson.put("languageCode", languageCode);
 
         Date date = new Date();
 
-        Map<String, String> parameters = new HashMap<>();
+        JSONObject parameters = new JSONObject();
         parameters.put("Solution", System.getenv("SOLUTION_ID"));
         parameters.put("UUID", System.getenv("UUID"));
         parameters.put("TimeStamp", date.toInstant().toString());
-        parameters.put("Data", json.toString());
+        parameters.put("Data", dataJson.toString());
         logger.info("Set all the parameters for the request: " + parameters.toString());
 
         return parameters;
     }
 
-    private static HttpURLConnection createConnection(Map<String, String> parameters) throws Exception {
+    private static HttpURLConnection createConnection(JSONObject parameters) throws Exception {
         URL url = new URL("https://metrics.awssolutionsbuilder.com/generic");
         HttpURLConnection connection = (HttpURLConnection) url.openConnection();
         connection.setRequestProperty("Content-Type", "application/json");
+        connection.setRequestProperty("Accept", "application/json");
         connection.setRequestMethod("POST");
         connection.setInstanceFollowRedirects(true);
-
         connection.setDoOutput(true);
-        DataOutputStream out = new DataOutputStream(connection.getOutputStream());
-        out.writeBytes(ParameterStringBuilder.getParamsString(parameters));
+
+        OutputStream out = connection.getOutputStream();
+        out.write(parameters.toString().getBytes());
         out.flush();
         out.close();